Pyro Prototype example
Smoothly running pyro example implementation. It exchanges the server's pyro object with the client. The client need to connect to the ip given to the server. It is just a prototype, but is exiting the correct way and uses threads as well as threading.Event. The server just accepts the pyro call an prints the type of action, depending on the method called by the client.