refactored generate List Handler
refactored the generate method from main method in it's own method. It can be used by importing examples.comparison.main.generate and afterwards may be used as generate(). It returns the generated list with items. Issue #4 is directly connected.

As an example I added the call to the createDict.py file.