changed listhandler
related to #4:

deleted listHandler.py from mergeDict directory - DO NOT COPY FILES ! This creates duplicates and is bad habit, please do not do it. Use the original implementation in util/listhandler.py with an import statement.

removed the get_dict method as it was for json-testing purpose only.

Added a to_json method to convert a copy of the dict into json.

Added a functionality to check if the given argument of add_or_override_entry actually is an Entry object. If not a TypeError is raised.

Changed methods to property decorators (setters and getters) at the Entry Class.

Refactored names of the attributes of Entry.

Renamed examples/comparison/main.py to listgenerator.py