related to #8, #6 & #5
copied observableInterface to util - example still exists.

Implemented an ObervableInterface implementation in manager.py. It holds the magical update method.

manager.py starts UDPServer which calls the handler in case of receiving packages. The handler calls the update method per callback. The UDPServer.serve_forever() is implemented as thread.

The sender.py holds the implementation of sending packages and waits for receiving packages within a second (to be discussed). Implemented as Thread.