closes #8
We have to test this commit in our big environment, as I dont know how the implementation reacts on flooding. Should not be that bad.

Multicasting functionality is implemented and documented. The list is successfully published and received by multicasting. The list is converted into json and for sending purpose is converted into bytes. On the receiving side, the received data is converted back into a string (bytes->string). The string in fact is the json representation of the list we received. There is already an implementation to get a python'ed dict out of this json-object (see listhandler).

As we have kickbacks here (I send data, I receive my own data), I defined that these data is dropped. This might me interesting for #4, because they will (most likely) kick the host, where the service is running, out of there list. So please DO Not kick the ip-address & Entry pair out of the list just because the timestamp is -1 or not updated at all. This is expected, because we do not want to process our own requests (sending the list to our self and merge it into our list is just pointless).

The whole functionality is documented, so read the docs - process the information - aks questions if something is unclear. Feel free to open issues if you find bugs.